Rename file for workflow purposes
The README.org -> README.md idea is good, but started getting in the way of Org-mode. Every time I want to generate the Agenda View, since the file had the '.org' extension, and also that file has a link to generate a pretty html which is located in a 3rd party repository in github, Emacs has to visit that link for every Agenda View generation; meaning a considerable amount of time comparing it to the rest of the operation.

Now it shouldn't visit that link every time I want to generate an Agenda View. More responsiveness.